initial commit of the daGloane Remote
The target is to make the daGloane qaudcopter controlable using a Android Smartphone.
For now to ease devolempent a common RC remote control will be used.
daGloane Remote is a stm32 discovery eval board and a bluetooth interface that can be connected to a RC remote control. It will read the PPM signal of the remote and transfer the channel information via bluetooth  to daGloane.